+++ jmbo_sitemap/__init__.py
@@ -36,6 +36,10 @@ def sitemap(request, sitemaps, section=None,
             raise Http404("Page %s empty" % page)
         except PageNotAnInteger:
             raise Http404("No page '%s'" % page)
+    # Google will not accept a sitemap with no urls. In such a case link back
+    # to the site.
+    if not urls:
+        urls = [{'location': '%s://%s' % (req_protocol, req_site.domain)}]
     return TemplateResponse(request, template_name, {'urlset': urls},
                             content_type=mimetype)