+++ CHANGELOG.rst
@@ -1,6 +1,11 @@
 Changelog
 =========
 
+0.1.2
+-----
+#. Omit external links from sitemap.
+#. Google will not accept a sitemap with no urls. In such a case link back to the site.
+
 0.1.1
 -----
 #. Use `django-setuptest>=0.1.4`.